+++ climaf/classes.py
@@ -10,10 +10,10 @@
 import re, string, copy, os.path
 
 import dataloc
-from period import init_period, cperiod
-from clogging import clogger, dedent
+from period    import init_period, cperiod
+from clogging  import clogger, dedent
 from netcdfbasics import fileHasVar, varsOfFile, timeLimits, model_id
-from decimal import Decimal
+from decimal   import Decimal
 
 #: Dictionary of declared projects (type is cproject)
 cprojects=dict()
@@ -126,6 +126,7 @@ class cproject():
         self.attributes_for_ensemble=['simulation']
         if 'ensemble' in kwargs :
             self.attributes_for_ensemble.extend(kwargs["ensemble"])
+
     def __repr__(self):
         return self.crs
     def crs2ds(self,crs) :
@@ -560,7 +561,7 @@ def fds(filename, simulation=None, variable=None, period=None, model=None) :
 
     For dataset attributes which are not provided, these defaults apply :
 
-    - simulation : the filename basename
+    - simulation : the filename basename (without suffix '.nc')
     - variable : the set of variables in the data file
     - period : the period actually covered by the data file (if it has time_bnds)
     - model : the 'model_id' attribute if it exists, otherwise : 'no_model'
@@ -582,7 +583,7 @@ def fds(filename, simulation=None, variable=None, period=None, model=None) :
         raise Climaf_Classes_Error("File %s does no exist"%filename)
     #
     if model is None : model=model_id(filename)
-    if simulation is None : simulation=os.path.basename(filename)
+    if simulation is None : simulation=os.path.basename(filename)[0:-3]
     #
     if variable is None :
         lvars=varsOfFile(filename)
@@ -718,6 +719,12 @@ def compare_trees(tree1,tree2,func,filter_on_operator=None) :
             return compare_trees(tree1.father,tree2.father,
                                  func,filter_on_operator)
 
+allow_errors_on_ds_call=False
+
+def allow_error_on_ds(allow=True) :
+    global allow_errors_on_ds_call
+    allow_errors_on_ds_call=allow
+    #print ('allow_errors_on_ds_call='+`allow_errors_on_ds_call`)
 
 def ds(*args,**kwargs) :
     """
@@ -743,17 +750,16 @@ def ds(*args,**kwargs) :
         except Climaf_Classes_Error: dataset=None
         if (dataset) : results.append(dataset)
     if len(results) > 1 :
-        e="CRS expressions %s is ambiguous among projects %s"%(crs,`cprojects.keys()`)
-        clogger.error(e)
-        raise Climaf_Classes_Error(e)
+        e="CRS expression %s is ambiguous among projects %s"%(crs,`cprojects.keys()`)
+        if allow_errors_on_ds_call : clogger.info(e)
+        else : raise Climaf_Classes_Error(e)
     elif len(results) == 0 :
-        e="CRS expressions %s is not valid for any project in %s"%(crs,`cprojects.keys()`)
-        raise Climaf_Classes_Error(e)
-        return None
+        e="CRS expression %s is not valid for any project in %s"%(crs,`cprojects.keys()`)
+        if allow_errors_on_ds_call : clogger.debug(e)
+        else : raise Climaf_Classes_Error(e)
     else : 
         rep=results[0]
-        if rep.project=='file' : 
-            rep.files=rep.kvp["path"]
+        if rep.project=='file' : rep.files=rep.kvp["path"]
         return rep
 
 def cfreqs(project,dic) :
@@ -837,28 +843,59 @@ def cmissing(project,missing,*kwargs) :
 
 class cpage(cobject):
     def __init__(self, fig_lines=None, widths=None, heights=None, 
-                  orientation="portrait", fig_trim=False, page_trim=False):
+                 orientation="portrait", fig_trim=True, page_trim=True, format="png",
+                 title="", x=0, y=26, ybox=50, pt=24,
+                 font="Times-New-Roman", gravity="North", background="white"): 
         """
         Builds a CliMAF cpage object, which represents an array of figures
 
         Args:
-         fig_line (a list of lists of figure objects or an ensemble of
-           figure objects): each sublist of 'fig_lines' represents a
-           line of figures
-         widths (list, optional): the list of figure widths, i.e. the
-           width of each column. By default, if fig_line is:
+        
+          fig_line (a list of lists of figure objects or an ensemble of figure objects) :
+           each sublist of 'fig_lines' represents a line of figures   
+          widths (list, optional) : the list of figure widths, i.e. the width of each 
+           column. By default, if fig_line is:
+
              - a list of lists:  spacing is even
              - an ensemble:  one column is used
-         heights (list, optional): the list of figure heights, i.e. the
+          heights (list, optional) : the list of figure heights, i.e. the
            height of each line. By default  spacing is even
-         orientation (str, optional): page's orientation, either 'portrait' 
+          orientation (str, optional) : page's orientation, either 'portrait' 
            (default) or 'landscape'
-         fig_trim (str, optional): to turn on/off triming for all figures.
+          fig_trim (logical, optional) : to turn on/off triming for all figures.
            It removes all the surrounding extra space of figures in the page,
-           either False (default) or True
-         page_trim (str, optional): to turn on/off triming for the page. It
-           removes all the surrounding extra space of the page, either False
-           (default) or True
+           either True (default) or False
+          page_trim (logical, optional) : to turn on/off triming for the page. It
+           removes all the surrounding extra space of the page, either True
+           (default) or False 
+          format (str, optional) : graphic output format, either 'png' (default)
+           or 'pdf'
+          title (str, optional) : append a label below or above (depending optional
+           argument 'gravity') figures in the page.
+
+        If title is activated:
+        
+            - x, y (int, optional) : annotate the page with text.
+              x is the offset towards the right from the upper left corner
+              of the page, while y is the offset upward or the bottom
+              according to the optional argument 'gravity' (i.e. 'South' or 'North'
+              respectively); CLiMAF default: x=0, y=26. For more details, see:
+              http://www.imagemagick.org/script/command-line-options.php?#annotate ;
+              where x and y correspond respectively to tx and ty
+              in ``-annotate {+-}tx{+-}ty text``
+            - ybox (int, optional) : width of the assigned box for title;
+              CLiMAF default: 50. For more details, see:
+              http://www.imagemagick.org/script/command-line-options.php?#splice
+            - pt (int, optional) : pt of the title; CLiMAF default: 24
+            - font (str, optional): set the font to use when creating title; CLiMAF
+              default: 'Times-New-Roman'. To print a complete list of fonts, use :
+              'convert -list font'
+            - gravity (str, optional) : the choosen direction specifies where to position
+              title; CLiMAF default: 'North'. For more details, see:
+              http://www.imagemagick.org/script/command-line-options.php?#gravity
+            - background (str, optional) : background color of the assigned box for
+              title; default: 'white'. To print a complete list of color names, use:
+              'convert -list color'
 
         Example:
 
@@ -866,9 +903,9 @@ class cpage(cobject):
         
           >>> fig=plot(tas_avg,title='title')
           >>> my_page=cpage([[None, fig],[fig, fig],[fig,fig]], widths=[0.2,0.8],
-          ... heights=[0.33,0.33,0.33], orientation='landscape', fig_trim=True, page_trim=True)
-
-        
+          ... heights=[0.33,0.33,0.33], orientation='landscape', fig_trim=False, page_trim=False,
+          ... format='pdf', title='Page title', x=10, y=20, ybox=45,
+          ... pt=20, font='Utopia', gravity='South', background='grey90' )
         """
         if fig_lines is None :
             raise Climaf_Classes_Error("fig_lines must be provided")
@@ -876,10 +913,20 @@ class cpage(cobject):
         self.orientation=orientation
         self.fig_trim=fig_trim
         self.page_trim=page_trim
-
+        self.format=format
+        self.title=title
+        self.x=x
+        self.y=y
+        self.ybox=ybox
+        self.pt=pt
+        self.font=font
+        self.gravity=gravity
+        self.background=background
+        if ( self.ybox < (self.y + self.pt) ) :
+            raise Climaf_Classes_Error("Title exceeds the assigned box: ybox<y+pt")
         if not isinstance(fig_lines,list) and not isinstance(fig_lines,cens) :
             raise Climaf_Classes_Error(
-                "fig_lines must be an ensemble or a list "
+                "fig_lines must be a CliMAF ensemble or a list "
                 "of lists (each representing a line of figures)")
         if isinstance(fig_lines,list) :
             if not widths :
@@ -936,14 +983,23 @@ class cpage(cobject):
                 if f : rep+=f.buildcrs(crsrewrite=crsrewrite)+","
                 else : rep+=`None`+","
             rep+=" ],"; 
-        rep+="],"+`self.widths`+","+`self.heights`+",orientation='"+self.orientation+\
-              "', fig_trim='%s', page_trim='%s')" %(self.fig_trim,self.page_trim)
+
+        if self.title is "" :
+            rep+=( "],"+`self.widths`+","+`self.heights`+", orientation='"+self.orientation+\
+                   "', fig_trim='%s', page_trim='%s', format='"+self.format+"')" )%(self.fig_trim,self.page_trim)
+        else:
+            rep+=( "],"+`self.widths`+","+`self.heights`+", orientation='"+self.orientation+\
+                   "', fig_trim='%s', page_trim='%s', format='"+self.format+"', title='"+self.title+\
+                   "', x=%d, y=%d, ybox=%d, pt=%d, font='"+self.font+\
+                   "', gravity='"+self.gravity+"', background='"+self.background+"')" )\
+                   %(self.fig_trim,self.page_trim,self.x,self.y,self.ybox,self.pt)
+            
         rep=rep.replace(",]","]")
         rep=rep.replace(", ]","]")
         
         return rep
 
-        
+
 def guess_projects(crs) :
     """
     Return the list of projects involved in the datasets involved in a 
@@ -954,8 +1010,8 @@ def guess_projects(crs) :
         Guess which is the project name for a dataset's crs, with minimum 
         assumption on the separator used in the project
         """
-        separators=[r'.',r'_',r'Â£',r'$',r'@',r'_',r'|',r'&',r"-",r"=",r"^",
-                    r";",r":",r"!",r'Â§',r'/',r'.',r'Ã¸',r'+',r'Â°']
+        separators=[r'.',r'_',r'£',r'$',r'@',r'_',r'|',r'&',r"-",r"=",r"^",
+                    r";",r":",r"!",r'§',r'/',r'.',r'ø',r'+',r'°']
         counts=dict()
         for sep in separators : counts[sep]=crs.count(sep)
         # Assume that the highest count gives the right separator
@@ -999,7 +1055,7 @@ class Climaf_Classes_Error(Exception):
         return `self.valeur`
 
 def test():
-#    clogger.basicConfig(level=clogger.DEBUG) #LV
+#    clogger.basicConfig(level=clogger.DEBUG) 
 #    clogger.basicConfig(format='"%(asctime)s [%(funcName)s: %(filename)s,%(lineno)d] %(message)s : %(levelname)s', level=clogger.DEBUG)
     cdef("project","CMIP5")
     #cdef("project","PR6")