+++ climaf/projects/ocmip5.py
@@ -4,7 +4,7 @@ This module declares how to access OCMIP5 data on Ciclad.
 
 Use attributes 'model' and 'frequency'
 
-Example of a path : /prodigfs/OCMIP5/OUTPUT/IPSL/IPSL-CM4/CTL/mon/CACO3/CACO3_IPSL_IPSL-CM4_CTL_1860-1869.nc
+Example of a path : /prodigfs/project/OCMIP5/OUTPUT/IPSL/IPSL-CM4/CTL/mon/CACO3/CACO3_IPSL_IPSL-CM4_CTL_1860-1869.nc
 
 
 Example ::
@@ -28,7 +28,7 @@ if onCiclad :
     # Declare which facets can be used for defining an ensemble
     #
     dataloc(project="OCMIP5", organization="generic",
-            url=['/prodigfs/OCMIP5/OUTPUT/*/${model}/${simulation}/${frequency}/'
+            url=['/prodigfs/project/OCMIP5/OUTPUT/*/${model}/${simulation}/${frequency}/'
                  '${variable}/${variable}_*_${model}_${simulation}_YYYY-YYYY.nc'])
     #
     cfreqs('OCMIP5',{'monthly':'mon' })