+++ examples/figarray.py
@@ -13,27 +13,46 @@ fig1=plot(tas,title="title")
 cfile(fig1)
 
 ###########################################################################
-# Define page1 as a figure array
-page1=cpage([[None, fig1],[fig1, fig1],[fig1,fig1]], widths=[0.2,0.8],heights=[0.33,0.33,0.33])
+# Define page1 as a figure array by trimming all the surrounding extra
+# space of figures (fig_trim=True by default) and also of the page
+# (page_trim=True by default)
+page1=cpage([[None, fig1],[fig1, fig1],[fig1,fig1]],
+            widths=[0.2,0.8],heights=[0.33,0.33,0.33])
 ###########################################################################
 # ask for display of page1
 cshow(page1)
 
 ###########################################################################
-# Define page2 as a figure array without specify widths and heights;
-# widths and heights will then be computed in a homogeneous way to fill the
-# page, i.e. for this example: widths=[0.5,0.5]; heights=[0.33,0.33,0.33]
-page2=cpage([[None, fig1],[fig1, fig1],[fig1,fig1]])
+# Same as page1 by adding a title and tuning some parameters for the title
 ###########################################################################
+page2=cpage([[None, fig1],[fig1, fig1],[fig1,fig1]],
+            widths=[0.2,0.8],heights=[0.33,0.33,0.33],title="Page title",
+            background="grey90",x=-300,y=26,pt=20,
+            font='Utopia',ybox=60)
 cshow(page2)
+ 
+###########################################################################
+# Same as page1 with a title but without tuning parameters for the title
+###########################################################################
+page3=cpage([[None, fig1],[fig1, fig1],[fig1,fig1]],
+            widths=[0.2,0.8],heights=[0.33,0.33,0.33],title="Page title")
+cshow(page3)
 
 ###########################################################################
-# Define page3 as a figure array by trimming all the surrounding extra
-# space of figures (fig_trim=True) and also of the page (page_trim=True) 
-page3=cpage([[None, fig1],[fig1, fig1],[fig1,fig1]],
-            fig_trim=True,page_trim=True)
+# Define page4 as a figure array without specify widths and heights;
+# widths and heights will then be computed in a homogeneous way to fill the
+# page, i.e. for this example: widths=[0.5,0.5]; heights=[0.33,0.33,0.33]
+page4=cpage([[None, fig1],[fig1, fig1],[fig1,fig1]])
 ###########################################################################
-cshow(page3)
+cshow(page4)
+
+###########################################################################
+# Define page5 as a figure array by not trimming all the surrounding extra
+# space of the page (page_trim=False) but of figures (fig_trim=True by default) 
+page5=cpage([[None, fig1],[fig1, fig1],[fig1,fig1]],
+            page_trim=False)
+###########################################################################
+cshow(page5)
 
 ###########################################################################
 # Define an ensemble with one single month of data per member
@@ -48,17 +67,17 @@ fig_ens=plot(ens,title="title")
 cshow(fig_ens) # will launch cshow once per member
 
 ###########################################################################
-# Define page4 as a figure array without specifying widths and heights;
+# Define page6 as a figure array without specifying widths and heights;
 # widths and heights will then be computed in an even way to fill the
 # page with one column, i.e. for this example: widths=[1.]; heights=[0.5,0.5]
-page4=cpage(fig_ens)
+page6=cpage(fig_ens)
 ###########################################################################
-cshow(page4)
+cshow(page6)
 
 ###########################################################################
-# Define page5 as a figure array by only specifying heights (in the case of
-# an ensemble : widths=[1.] by default) and by trimming all the surrounding
-# extra space of figures (fig_trim=True) and also of the page (page_trim=True) 
-page5=cpage(fig_ens, heights=[0.8,0.2],fig_trim=True, page_trim=True)
+# Define page7 as a figure array by only specifying heights (in the case of an
+# ensemble : widths=[1.] by default) and by not trimming all the surrounding extra
+# space of the page (page_trim=False) but of figures (fig_trim=True by default) 
+page7=cpage(fig_ens, heights=[0.8,0.2], page_trim=False)
 ###########################################################################
-cshow(page5)
+cshow(page7)